requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-dotstar==2.2.12
adafruit-circuitpython-pixelbuf==2.0.5
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.1
args==0.1.0
binho-host-adapter==0.1.6
blinker==1.8.2
click==8.1.7
clint==0.5.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
lgpio==0.2.2.0
MarkupSafe==2.1.5
PyDMXControl==2.0.0
pyftdi==0.55.4
pygame==2.6.0
pyserial==3.5
pyusb==1.2.1
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
udmx-pyusb==2.0.0
Werkzeug==3.0.3
